+++ packages/ui/Routable/package.json
@@ -1,0 +1,3 @@
+{
+	"main": "Routable.js"
+}
